refactor: use unstable cache where applicable
