+++ src/app/proposals/actions/proposalsAction.ts
@@ -7,6 +7,7 @@ import { GovernorAbi } from '@/lib/abis/Governor'
 import { unstable_cache } from 'next/cache'
 import { gql as apolloGQL } from '@apollo/client'
 import { daoClient } from '@/shared/components/ApolloClient'
+import { AVERAGE_BLOCKTIME } from '@/lib/constants'
 
 const fetchProposalSharedDetails = async () => {
   // Proposal Threshold (from governor)
@@ -107,6 +108,6 @@ async function fetchProposals() {
 }
 
 export const getCachedProposals = unstable_cache(fetchProposals, ['cached_proposals'], {
-  revalidate: 60, // Every 60 seconds
+  revalidate: AVERAGE_BLOCKTIME,
   tags: ['cached_proposals'],
 })
